First wave of circular dependencies fix
