+++ front/components/data_source/ZendeskOAuthExtraConfig.tsx
@@ -2,7 +2,7 @@ import { Input } from "@dust-tt/sparkle";
 import { useEffect } from "react";
 
 import type { ConnectorOauthExtraConfigProps } from "@app/lib/connector_providers_ui";
-import { isValidZendeskSubdomain } from "@app/types";
+import { isValidZendeskSubdomain } from "@app/types/oauth/lib";
 
 export function ZendeskOAuthExtraConfig({
   extraConfig,
