+++ front/components/editor/extensions/skill_builder/KnowledgeNode.tsx
@@ -1,53 +1,15 @@
 import { mergeAttributes, Node } from "@tiptap/core";
 import { ReactNodeViewRenderer } from "@tiptap/react";
 
-import { KnowledgeNodeView } from "@app/components/editor/extensions/skill_builder/KnowledgeNodeView";
-import type { DataSourceViewContentNode } from "@app/types/data_source_view";
-
-// Minimal data from serialization.
-export interface BaseKnowledgeItem {
-  dataSourceViewId: string;
-  hasChildren: boolean;
-  label: string;
-  nodeId: string;
-  spaceId: string;
-}
-
-// Fresh selection from search with complete node data.
-export interface FullKnowledgeItem extends BaseKnowledgeItem {
-  node: DataSourceViewContentNode;
-}
-
-export type KnowledgeItem = BaseKnowledgeItem | FullKnowledgeItem;
-
-export function isFullKnowledgeItem(
-  item: KnowledgeItem
-): item is FullKnowledgeItem {
-  return "node" in item && item.node !== undefined;
-}
-
-/**
- * Computes whether a node has children, with special handling for Notion.
- * For Notion: pages and databases can have children even if they're currently empty.
- * For others: uses expandable field or node type.
- */
-export function computeHasChildren(node: DataSourceViewContentNode): boolean {
-  const isNotion =
-    node.dataSourceView.dataSource.connectorProvider === "notion";
-
-  if (isNotion) {
-    // In Notion, pages (documents) and databases (tables) can have children.
-    // Folders always can have children (though Notion doesn't actually use folders).
-    return (
-      node.type === "folder" ||
-      node.type === "document" ||
-      node.type === "table"
-    );
-  }
-
-  // For non-Notion sources, use the childrenCount field.
-  return node.childrenCount > 0;
-}
+import type {
+  BaseKnowledgeItem,
+  KnowledgeItem,
+} from "@app/components/editor/extensions/skill_builder/KnowledgeNodeView";
+import {
+  computeHasChildren,
+  isFullKnowledgeItem,
+  KnowledgeNodeView,
+} from "@app/components/editor/extensions/skill_builder/KnowledgeNodeView";
 
 export interface KnowledgeNodeAttributes {
   selectedItems: KnowledgeItem[];
