+++ front/components/editor/extensions/skill_builder/KnowledgeNodeView.tsx
@@ -23,15 +23,7 @@ import {
   KnowledgeChip,
   KnowledgeErrorChip,
 } from "@app/components/editor/extensions/skill_builder/KnowledgeChip";
-import type {
-  FullKnowledgeItem,
-  KnowledgeItem,
-  KnowledgeNodeAttributes,
-} from "@app/components/editor/extensions/skill_builder/KnowledgeNode";
-import {
-  computeHasChildren,
-  isFullKnowledgeItem,
-} from "@app/components/editor/extensions/skill_builder/KnowledgeNode";
+import type { KnowledgeNodeAttributes } from "@app/components/editor/extensions/skill_builder/KnowledgeNode";
 import { getConnectorProviderLogoWithFallback } from "@app/lib/connector_providers_ui";
 import {
   getLocationForDataSourceViewContentNodeWithSpace,
@@ -43,6 +35,52 @@ import { useUnifiedSearch } from "@app/lib/swr/search";
 import { useSpaceDataSourceView, useSpaces } from "@app/lib/swr/spaces";
 import type { LightWorkspaceType } from "@app/types";
 import { removeNulls } from "@app/types";
+import type { DataSourceViewContentNode } from "@app/types/data_source_view";
+
+// Minimal data from serialization.
+export interface BaseKnowledgeItem {
+  dataSourceViewId: string;
+  hasChildren: boolean;
+  label: string;
+  nodeId: string;
+  spaceId: string;
+}
+
+// Fresh selection from search with complete node data.
+export interface FullKnowledgeItem extends BaseKnowledgeItem {
+  node: DataSourceViewContentNode;
+}
+
+export type KnowledgeItem = BaseKnowledgeItem | FullKnowledgeItem;
+
+export function isFullKnowledgeItem(
+  item: KnowledgeItem
+): item is FullKnowledgeItem {
+  return "node" in item && item.node !== undefined;
+}
+
+/**
+ * Computes whether a node has children, with special handling for Notion.
+ * For Notion: pages and databases can have children even if they're currently empty.
+ * For others: uses expandable field or node type.
+ */
+export function computeHasChildren(node: DataSourceViewContentNode): boolean {
+  const isNotion =
+    node.dataSourceView.dataSource.connectorProvider === "notion";
+
+  if (isNotion) {
+    // In Notion, pages (documents) and databases (tables) can have children.
+    // Folders always can have children (though Notion doesn't actually use folders).
+    return (
+      node.type === "folder" ||
+      node.type === "document" ||
+      node.type === "table"
+    );
+  }
+
+  // For non-Notion sources, use the childrenCount field.
+  return node.childrenCount > 0;
+}
 
 interface KnowledgeDisplayProps {
   item: KnowledgeItem;
