+++ front/lib/api/oauth/providers/zendesk.ts
@@ -1,5 +1,6 @@
 import type { ParsedUrlQuery } from "querystring";
 
+import { isValidZendeskSubdomain } from "@app/lib/api/actions/servers/zendesk/types";
 import config from "@app/lib/api/config";
 import type { BaseOAuthStrategyProvider } from "@app/lib/api/oauth/providers/base_oauth_stragegy_provider";
 import {
@@ -7,7 +8,6 @@ import {
   getStringFromQuery,
 } from "@app/lib/api/oauth/utils";
 import type { ExtraConfigType } from "@app/pages/w/[wId]/oauth/[provider]/setup";
-import { isValidZendeskSubdomain } from "@app/types";
 import type { OAuthConnectionType, OAuthUseCase } from "@app/types/oauth/lib";
 
 export class ZendeskOAuthProvider implements BaseOAuthStrategyProvider {
